+++ CHANGES.md
@@ -1,5 +1,11 @@
 # load-stylesheets
 
+## 0.2.0
+
+- Enhancement: Allow favicon loading
+- Enhancement: Add `after` option
+- Testing: Fix single stylesheet test (only one color should be applied)
+
 ## 0.1.1
 
 - npm fix: Ensure `main` is correct